requirements_dev.txt
bump2version
flake8
GitPython
interrogate
mypy
pandas
pytest
pytest-cov
pytest-mock
PyYAML
setuptools
wheel
